fix: keep contract pagination tests watcher-free
